Move OverlayHeaderTabControl to correct namespace
